Separate Webpack configs & add ssr config
